Removed excessive divs and h2 moved out of divs
The h2 tag being inside the div restricted the use of flex. The first
solution was to add an extra div to separate the h2 & the illustrations,
but this resulted in too many nested divs. The h2 tag has been moved out
of the information-container div to allow the use of flex. Some styling
has been added to the illustrations & information-container divs